Fix OneDrive file listing API usage (#145)
OneDrive recently changed such that the `$filter` OData query parameter
can no longer be used when getting children of a directory.

So, instead of querying the folders and images separately, we'll now
query for all children and then filter based on if the `image` or
`folder` property is available for each.

Also: bumped dependencies, reduced the number of items being queried at
once.